Yet another nlab scraper modification
I realized that there are two standards of formatting the table of
contents. One labels it 'Contents' and the other labels it with the
article title. To handle this, I modified my scraper again to first
just look for a table of contents div, and skip past it if it found
one, and otherwise to just look at divs from the beginning. This seems
to work for now